copy_api_tables reworked
